Fixed updateUser to handle currentUser correctly.
Current user _id was an object, so it needed to be changed to a string in order to compare it to id. _currentUser needed to also obtain the _id of the currentUser. This makes it so changes to the profile will be visible when returning back to the profile from another screen.